Refactor app structure and modularize services
Reorganized the folder structure across services to comply with modular standards, improving maintainability. Deprecated and removed redundant IdentityServiceController, and migrated initialization logic to a shared microservice bootstrap function. Adjusted imports and references to reflect the new structure.
