+++ api/apps/feature-flag-service/src/app/feature-flag-service.module.ts
@@ -1,14 +1,15 @@
+import {PrismaModule} from '@app/database';
 import {ConfigifyModule} from '@jdevel/configify';
 import {Module} from '@nestjs/common';
 import {ConfigModule} from '@nestjs/config';
 import {EventEmitterModule} from '@nestjs/event-emitter';
 
-import {EmbedScriptsModule} from './embed-scripts/embed-scripts.module';
-import {IssuesModule} from './issues/issues.module';
-import {ProductsModule} from './products/products.module';
+import {PrismaClient as FeatureFlagPrismaClient} from '../../generated/client';
 
 @Module({
   imports: [
+    // TODO: Move to individual modules
+    PrismaModule.register({client: FeatureFlagPrismaClient}),
     ConfigModule.forRoot({
       isGlobal: true,
     }),
@@ -16,9 +17,6 @@ import {ProductsModule} from './products/products.module';
     EventEmitterModule.forRoot({
       global: true,
     }),
-    EmbedScriptsModule,
-    IssuesModule,
-    ProductsModule,
   ],
 })
-export class AppModule {}
+export class FeatureFlagServiceModule {}
