+++ api/apps/feedback-service/src/app/embed-scripts/controllers/embed-scripts/embed-scripts.controller.spec.ts
@@ -1,7 +1,7 @@
 import {Mocked, TestBed} from '@suites/unit';
 
 import {EmbedScriptsController} from './embed-scripts.controller';
-import {ProjectsService} from '../../../../../../identity-service/src/projects/services/projects/projects.service';
+import {ProjectsService} from '../../../../../../identity-service/src/app/projects/services/projects/projects.service';
 
 describe('EmbedScriptsController', () => {
   let controller: EmbedScriptsController;
