+++ api/apps/feedback-service/src/app/embed-scripts/embed-scripts.module.ts
@@ -1,7 +1,7 @@
 import {Module} from '@nestjs/common';
 
 import {EmbedScriptsController} from './controllers/embed-scripts/embed-scripts.controller';
-import {ProjectsModule} from '../../../../identity-service/src/projects/projects.module';
+import {ProjectsModule} from '../../../../identity-service/src/app/projects/projects.module';
 
 @Module({
   imports: [ProjectsModule],
