+++ api/apps/feedback-service/src/app/products/products.module.ts
@@ -4,7 +4,7 @@ import {Logger, Module} from '@nestjs/common';
 import {ProductsController} from './controllers/products/products.controller';
 import {ProductsRepositoryService} from './repositories/products-repository/products-repository.service';
 import {ProductsService} from './services/products/products.service';
-import {ProjectsModule} from '../../../../identity-service/src/projects/projects.module';
+import {ProjectsModule} from '../../../../identity-service/src/app/projects/projects.module';
 import {PrismaClient as FeedbackPrismaClient} from '../../../generated/client';
 
 @Module({
