+++ api/apps/identity-service/src/app/clients/dto/IsSubdomainAvailable.dto.ts
