+++ api/apps/identity-service/src/app/clients/repositories/clients-repository/clients-repository.service.it-spec.ts
@@ -9,8 +9,8 @@ import {StartedPostgreSqlContainer} from '@testcontainers/postgresql';
 import {Client} from 'pg';
 
 import {ClientsRepositoryService} from './clients-repository.service';
+import {PaymentsService} from '../../../../../../payments-service/src/app/services/payments/payments.service';
 import {PaymentsModule} from '../../../../../payments-service/src/payments.module';
-import {PaymentsService} from '../../../../../payments-service/src/services/payments/payments.service';
 import {UsersRepositoryService} from '../../../users/repositories/users-repository/users-repository.service';
 import {UsersModule} from '../../../users/users.module';
