+++ api/apps/identity-service/src/app/users/services/authenticated-users/authenticated-users.service.spec.ts
