--- api/apps/identity-service/src/identity-service.service.ts
@@ -1,8 +1,0 @@
-import {Injectable} from '@nestjs/common';
-
-@Injectable()
-export class IdentityServiceService {
-  getHello(): string {
-    return 'Hello World!';
-  }
-}
