+++ api/apps/payments-service/src/main.ts
@@ -1,9 +1,15 @@
-import {NestFactory} from '@nestjs/core';
+import {feedbackServiceConstants} from '@app/constants';
+import {bootstrapMicroservice} from '@app/init';
+import {configDotenv} from 'dotenv';
 
-import {PaymentsServiceModule} from './payments-service.module';
+import {PaymentsServiceModule} from './app/payments-service.module';
 
-async function bootstrap() {
-  const app = await NestFactory.create(PaymentsServiceModule);
-  await app.listen(process.env.port ?? 3000);
-}
-bootstrap();
+configDotenv();
+
+bootstrapMicroservice(
+  PaymentsServiceModule,
+  [...(process.env.RABBIT_MQ_URLS?.split(',') ?? [])],
+  feedbackServiceConstants.queueName,
+  'feedbackUrl',
+  './apps/payments-service/prisma/schema.prisma',
+).catch((error) => console.error(error));
