+++ api/apps/payments-service/test/app.e2e-spec.ts
@@ -2,7 +2,7 @@ import {INestApplication} from '@nestjs/common';
 import {Test, TestingModule} from '@nestjs/testing';
 import * as request from 'supertest';
 
-import {PaymentsServiceModule} from './../src/payments-service.module';
+import {PaymentsServiceModule} from '../src/app/payments-service.module';
 
 describe('PaymentsServiceController (e2e)', () => {
   let app: INestApplication;
